pyproject.toml
[project]
name = "anki-generator"
version = "0.1.0"
description = "Universal CLI Anki deck generator using Claude API"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
]

[project.scripts]
anki-gen = "src.main:cli"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["src"]
